moves WithLocalFiles to common storages configuration
