+++ dlt/destinations/impl/qdrant/configuration.py
@@ -10,7 +10,8 @@ from dlt.common.configuration.specs.base_configuration import (
 )
 from dlt.common.destination.client import DestinationClientDwhConfiguration
 from dlt.common.utils import digest128
-from dlt.destinations.configuration import WithLocalFiles
+from dlt.common.storages.configuration import WithLocalFiles
+
 from dlt.destinations.impl.qdrant.exceptions import InvalidInMemoryQdrantCredentials
 from dlt.destinations.impl.qdrant.warnings import location_on_credentials_deprecated
