+++ tests/load/lancedb/test_config.py
@@ -68,9 +68,9 @@ def test_lancedb_follows_local_dir() -> None:
     # check pipeline name
     pipeline = dlt.pipeline("test_lancedb_follows_local_dir")
     c = resolve_configuration(
-        LanceDBClientConfiguration()
-        ._bind_dataset_name(dataset_name="test_dataset")
-        ._bind_pipeline(pipeline)
+        pipeline._bind_local_files(
+            LanceDBClientConfiguration()._bind_dataset_name(dataset_name="test_dataset")
+        )
     )
     db_path = os.path.join(local_dir, "test_lancedb_follows_local_dir.lancedb")
     assert c.lance_uri.endswith(db_path)
